Abstract Types

ChemistryFeaturization.AbstractType.AbstractAtoms — Type

AbstractAtoms

All types defined for different representations of atomic systems (different Atoms types) must be a subtype of AbstractAtoms.

ChemistryFeaturization.AbstractType.AbstractCodec — Type

AbstractCodec

All codecs defined for different encoding-decoding schemes must be a subtype of AbstractCodec.

ChemistryFeaturization.AbstractType.AbstractFeatureDescriptor — Type

AbstractFeatureDescriptor

All FeatureDescriptors defined for different types of features must be a subtype of AbstractFeatureDescriptor.

ChemistryFeaturization.AbstractType.AbstractFeaturization — Type

AbstractFeaturization

All types defined for different featurizations must be a subtype of AbstractFeaturization.