Atoms Objects

ChemistryFeaturization.Atoms — Module

Atoms

Atoms objects characterize structural representations of molecular structures. These representations can either be standardized forms (like graph representation using an adjacency matrix, point cloud representation, etc), or be custom user-defined representations too.

Atoms objects (terminology borrowed from ASE) store information about the structure of a molecule, crystal, etc. as well as, optionally, encoded features and the featurization used to encode them. The parent abstract type is AbstractAtoms.

AtomGraph

The AtomGraph type is used to store atomic graph representations. It can also be visualized using some customized formatting with the GraphPlot package.

ChemistryFeaturization.Atoms.AtomGraph — Type

AtomGraph

A type representing an atomic structure as a graph (gr).

Fields

graph::SimpleWeightedGraph{<:Integer,<:Real}: the graph representing the structure. See build_graph for more on generating the weights.

elements::Vector{String}: list of elemental symbols corresponding to each node of the graph

laplacian::Matrix{<:Real}: Normalized graph Laplacian matrix, stored to speed up convolution operations by avoiding recomputing it every pass.

id::String: Optional, an identifier, e.g. to correspond with tags/labels of an imported dataset.

ChemistryFeaturization.Atoms.AtomGraph — Method

AtomGraph(input_file_path, id = splitext(input_file_path)[begin]; output_file_path = nothing, featurization = nothing, overwrite_file = false, use_voronoi = false, cutoff_radius = 8.0, max_num_nbr = 12, dist_decay_func = inverse_square, normalize_weights = true)

Construct an AtomGraph object from a structure file.

Required Arguments

input_file_path::String: path to file containing structure (must be readable by ASE.io.read)

Optional Arguments

id::String: ID associated with structure (e.g. identifier from online database). Defaults to name of input file if undefined.

output_file_path = nothing: If provided, structure will be serialized to file at this location

overwrite_file::Bool = false: whether to overwrite an existing file at output_file_path

use_voronoi::Bool = false: Whether to build neighbor lists using Voronoi decompositions

cutoff_radius::Real = 8.0: If not using Voronoi neighbor lists, longest allowable distance to a neighbor, in Angstroms

max_num_nbr::Integer = 12: If not using Voronoi neighbor lists, largest allowable number of neighbors

dist_decay_func = inverse_square: Function by which to assign edge weights according to distance between neighbors

normalize_weights::Bool = true: Whether to normalize weights such that the largest is 1.0

Note

max_num_nbr is a "soft" limit – if multiple neighbors are at the same distance, the full neighbor list may be longer.

ChemistryFeaturization.Atoms.visualize — Function

Visualize a given graph.