Codecs

ChemistryFeaturization.Codec — Module

Codec

Codecs provide the encoding-decoding scheme in a flexible way. Codecs are titled according to the general attributes that the codec scheme would require, and NOT the actual encoding/decoding function itself.

This makes things customizable, and allows plug-and-play behaviour with different variants of the same codec scheme, or for strikingly similar codec schemes.

ChemistryFeaturization.Codec.OneHotOneCold — Type

OneHotOneCold(categorical, bins)

AbstractCodec type which uses a dummy variable (as defined in statistical literature), i.e., which employs one-hot encoding and a one-cold decoding scheme.

ChemistryFeaturization.Codec.DirectCodec — Type

DirectCodec

Codec type whose encoding function is some constant * the identity function.