Featurization

ChemistryFeaturization.Featurization — Module

Featurization

A Featurization collectively stores a set of FeatureDescriptors (and any other supporting attributes that it may further require), and defines how to featurize an Atoms object by using the encoding schemes defined by the FeatureDescriptors stored.

Featurization defines how encoded values of different FeatureDescriptors can be collated and modified.

The implementation for a featurization must be as standardized and generic as possible, so that FeatureDescriptors that may be built using/use different datasets can be built and used easily.

ChemistryFeaturization.Featurization.GraphNodeFeaturization — Type

GraphNodeFeaturization(encoded_atom_features::Vector{AtomFeatureDescriptor})
GraphNodeFeaturization(feature_names, lookup_table, nbins, logspaced, categorical)

A featurization for AtomGraph objects that encodes features associated with each node. Contains a collection of AbstractAtomFeatureDescriptor objects, and can be initialized by passing those, or by passing parameters for constructing them.

Required Arguments

feature_names::Vector{String}: Names of each feature

lookup_table::DataFrame (optional): source of data for features, if not included in built-in atom_data_df

Keyword Arguments

nbins::Union{Vector{Integer},Integer}: Number of bins to use for one-hot encoding of continuous-valued features. Will be ignored for categorical features.

logspaced::Union{Vector{Bool},Bool}: Whether to logarithmically space the bins

categorical::Union{Vector{Bool},Bool}: Whether each feature is categorical or continous-valued.