Featurized Atoms

A FeaturizedAtoms object is a container for an Atoms object, a compatible Featurization object, and the resulting encoded features.
The type is parameterized by the type of atoms object and the featurization type, e.g. FeaturizedAtoms{AtomGraph,GraphNodeFeaturization}.

This is intended to be the type that is directly fed into a Chemellia model, so the format of encoded_features should be whatever the associated model requires, and not necessarily human-readable (though interpretable through the decode function!).

ChemistryFeaturization.FeaturizedAtoms — Type

FeaturizedAtoms

Container object for an Atoms object, a Featurization, and the resulting encoded_features from applying the featurization to the atoms.

Fields

atoms: Atoms object to be featurized

featurization: Featurization scheme meant to be used for featurizing atoms

encoded_features: The result of featurizing atoms using featurization

Note

encoded_features will NOT change for a given atoms-featurization pair.