Utilities

Some helper/utility functions that aren't documented elsewhere...

ChemistryFeaturization.Utils.ElementFeatureUtils.default_categorical

ChemistryFeaturization.Utils.ElementFeatureUtils.default_log

ChemistryFeaturization.Utils.ElementFeatureUtils.fea_minmax

ChemistryFeaturization.Utils.ElementFeatureUtils.get_bins

ChemistryFeaturization.Utils.ElementFeatureUtils.get_param_vec

ChemistryFeaturization.Utils.ElementFeatureUtils.onecold_decoder

ChemistryFeaturization.Utils.ElementFeatureUtils.onehot_lookup_encoder

ChemistryFeaturization.Utils.GraphBuilding.build_graph

ChemistryFeaturization.Utils.GraphBuilding.weights_cutoff

ChemistryFeaturization.Utils.GraphBuilding.weights_voronoi

ChemistryFeaturization.Utils.ElementFeatureUtils.default_categorical — Function

default_categorical(feature_name, lookup_table = atom_data_df)

Determine if a feature should be treated as categorical or continuous-valued.

If the value type is not a number, always returns true. If it is, checks whether it is in the built-in list of categorical features.

TODO: possibly add behavior where it will default to categorical if there is below some threshold number of discrete values?

ChemistryFeaturization.Utils.ElementFeatureUtils.default_log — Function

default_log(feature_name, lookup_table = atom_data_df; threshold = oom_threshold_log)

Determine whether a continuous-valued feature should have logarithmically spaced bins.

Operates by finding the minimum and maximum values the feature can take on and comparing their ratio to a specified order-of-magnitude threshold that defaults to a package constant if not provided.

ChemistryFeaturization.Utils.ElementFeatureUtils.fea_minmax — Function

Compute the minimum and maximum possible values of a feature.

ChemistryFeaturization.Utils.ElementFeatureUtils.get_bins — Function

Helper function for encoder and decoder...(nbins is ignored for categorical=true)

ChemistryFeaturization.Utils.ElementFeatureUtils.get_param_vec — Method

Little helper function to check that the logspace/categorical vector/boolean is appropriate and convert it to a vector as needed.

ChemistryFeaturization.Utils.ElementFeatureUtils.onecold_decoder — Function

Function for decoding onehot-style vectors based on values in a lookup table. Intended to be used as an decode_f for ElementFeatureDescriptor objects. See source code of convenience constructor for ElementFeatureDescriptor for more details.

ChemistryFeaturization.Utils.ElementFeatureUtils.onehot_lookup_encoder — Function

Function for encoding onehot-style vectors based on values in a lookup table. Intended to be used as an encode_f for ElementFeatureDescriptor objects. See source code of convenience constructor for ElementFeatureDescriptor for more details.

ChemistryFeaturization.Utils.GraphBuilding.build_graph — Method

Function to build graph from a file storing a crystal structure (currently supports anything ase.io.read can read in). Returns an AtomGraph object.

Arguments

Required Arguments

file_path::String: Path to ASE-readable file containing a molecule/crystal structure

Keyword Arguments

normalize_weights::Bool=true: Whether to rescale graph weights such that the maximum value is 1.0 (recommended)

use_voronoi::bool: if true, use Voronoi method for neighbor lists, if false use cutoff method

(The rest of these parameters are only used if use_voronoi==false)

cutoff_radius::Real=8.0: cutoff radius for atoms to be considered neighbors (in angstroms)

max_num_nbr::Integer=12: maximum number of neighbors to include (even if more fall within cutoff radius)

dist_decay_func::Function=inverse_square: function to determine falloff of graph edge weights with neighbor distance

ChemistryFeaturization.Utils.GraphBuilding.weights_cutoff — Method

Build graph using neighbor number cutoff method adapted from original CGCNN. Note that max_num_nbr is a "soft" max, in that if there are more of the same distance as the last, all of those will be added.

ChemistryFeaturization.Utils.GraphBuilding.weights_voronoi — Method

Build graph using neighbors from faces of Voronoi polyedra and weights from areas. Based on the approach from https://github.com/ulissigroup/uncertaintybenchmarking/blob/aabb407807e35b5fd6ad06b14b440609ae09e6ef/BNN/datapyro.py#L268