Atoms Objects

Atoms objects (terminology borrowed from ASE) store information about the structure of a molecule, crystal, etc. as well as, optionally, encoded features and the featurization used to encode them. The parent abstract type is AbstractAtoms.

AtomGraph

The AtomGraph type is used to store atomic graph representations.

ChemistryFeaturization.Atoms.AtomGraph — Type

AtomGraph

A type representing an atomic structure as a graph (gr).

Fields

graph::SimpleWeightedGraph{<:Integer,<:Real}: the graph representing the structure. See build_graph for more on generating the weights.

elements::Vector{String}: list of elemental symbols corresponding to each node of the graph

laplacian::Matrix{<:Real}: Normalized graph Laplacian matrix, stored to speed up convolution operations by avoiding recomputing it every pass.

id::String: Optional, an identifier, e.g. to correspond with tags/labels of an imported dataset.

WeaveMol